+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Next
 
+- Add functionality to align piles by the grid via the context menu
 - Add a new property called [`pileCellAlign`](DOCS.md#pilingsetproperty-value) to define where the piles are aligned to, which accepts the following values:
   - 'topleft'
   - 'topRight'
@@ -34,6 +35,7 @@
   - `pileDrag`
   - `pileDrop`
 - Change the grid layout properties from a nested object called `grid` to the following individual properties that you can set with [`.set()`](DOCS.md#pilingsetproperty-value):
+
   - `itemSize`
   - `itemPadding`
   - `columns`
@@ -41,6 +43,7 @@
   - `cellAspectRatio`
 
   Their precedence is as follows:
+
   - `columns` < `itemSize`
   - `cellAspectRatio` < `rowHeight`